+++ src/ui/widget/mod.rs
@@ -96,7 +96,7 @@ pub trait Widget: Any {
         self
     }
     /// Set the font size of the inside text
-    fn set_label_size(self, size: u32) -> Self
+    fn set_label_size(self, size: f32) -> Self
     where
         Self: Sized,
     {
@@ -112,7 +112,7 @@ pub trait Widget: Any {
         self
     }
     /// Set the x-axis position of the widget
-    fn set_x(self, x: u32) -> Self
+    fn set_x(self, x: f64) -> Self
     where
         Self: Sized,
     {
@@ -120,7 +120,7 @@ pub trait Widget: Any {
         self
     }
     /// Set the y-axis position of the widget
-    fn set_y(self, y: u32) -> Self
+    fn set_y(self, y: f64) -> Self
     where
         Self: Sized,
     {
@@ -128,7 +128,7 @@ pub trait Widget: Any {
         self
     }
     /// Set the height dimension of the widget
-    fn set_height(self, height: u32) -> Self
+    fn set_height(self, height: f64) -> Self
     where
         Self: Sized,
     {
@@ -136,7 +136,7 @@ pub trait Widget: Any {
         self
     }
     /// Set the width dimension of the widget
-    fn set_width(self, width: u32) -> Self
+    fn set_width(self, width: f64) -> Self
     where
         Self: Sized,
     {
